requirements.txt
altair==5.2.0
annotated-types==0.7.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-ninja==1.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
executing==2.0.1
extra-streamlit-components==0.1.60
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
ipdb==0.13.13
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
numpy==1.26.3
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyarrow==15.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.25
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.31.0
streamlit-authenticator==0.2.3
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
zipp==3.17.0
